% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to comp
}
\examples{
\dontrun{
data(data); attach(data)
tm1 <- microbenchmark::microbenchmark(
rnR = gibbsR(100,10),
...
}
}
