% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curve_data}
\alias{curve_data}
\title{Example data of the curve fillter for a worm}
\format{
A Matrix with 43,288 rows times 100 columns.
}
\source{
\url{http://www.wenlab.org/}
}
\usage{
data(curve_data)
}
\description{
Each line represents the curvature of 100 segments of the body at a moment
Column i represents the curvature of the body of the ith segment over time.
}
\seealso{
\link{time_Elapsed},\link{w}
}
\keyword{datasets}
