% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedding_delay.R
\name{Embedding_delay}
\alias{Embedding_delay}
\title{Embedding the data after PCA dimensionality reduction}
\usage{
Embedding_delay(X, K_mode_dimension = 5, K_delay = 12)
}
\arguments{
\item{X}{a matrix with T*N size}

\item{K_mode_dimension}{is the dimensionality of PCA dimensionality reduction}

\item{K_delay}{is the delay time}
}
\value{
a matrix
}
\description{
Embedding the data after PCA dimensionality reduction
}
\examples{
\dontrun{
data("curve_data","time_Elapsed","w")
## you can use your data to Proceed to the steps below
rec <- Downsample(time_Elapsed,curve_data, 16)
time_downsampled <- rec$time_downsampled
data_downsampled <- rec$data_downsampled
ev <- Get_Matrix_Eigens(data_downsampled)
eigenvalues <- ev$values
eigenvectors <- ev$vectors
PCAmode <- 5
K_delay <- 12
Xeigenvector <- eigenvectors[ ,1:PCAmode]
X_PCA <- data_downsampled \%*\% Xeigenvector
test_CurveFilter_embeddingdelay <- Embedding_delay(X_PCA,PCAmode,K_delay)
X <- test_CurveFilter_embeddingdelay \%*\% w
## Plot the projection
plot(X[,3],X[,4],type = "l")
plot(X[,6],X[,7],type = "l")
}
}
