% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21049.R
\docType{package}
\name{foo}
\alias{foo}
\title{foo: A package for Calculate the phase space projection.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions…
}

