% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Matrix_Dimension.R
\name{Get_Matrix_Dimension}
\alias{Get_Matrix_Dimension}
\title{A function that Calculates the dimension of the matrix}
\usage{
Get_Matrix_Dimension(X)
}
\arguments{
\item{X}{a matrix with  T*N size}
}
\value{
a number, the dimension of the matrix X
}
\description{
Returns the dimension of the matrix X
}
\examples{
\dontrun{
X <- matrix(rnorm(200*100),200,100)
ev <- Get_Matrix_Dimension(X)
}
}
