% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Matrix_Eigens.R
\name{Get_Matrix_Eigens}
\alias{Get_Matrix_Eigens}
\title{A function that calculates matrix eigenvalues and eigenvectors}
\usage{
Get_Matrix_Eigens(X)
}
\arguments{
\item{X}{a matrix with T*N size}
}
\value{
a list with eigenvalues and eigenvectors
}
\description{
Returns a list of eigenvalues and eigenvectors for a matrix
}
\examples{
\dontrun{
X <- matrix(rnorm(200 * 100), 200, 100)
ev <- Get_Matrix_Eigens(X)
eigenvalues <- ev$values
eigenvectors <- ev$vectors
}
}
