% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin)
}
\arguments{
\item{N}{the length of the chain}

\item{thin}{the number of Binomial(thin, y)}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(1000,50)
plot(rnC[1,],type='l')
plot(rnC[2,],type='l')
}
}
