% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsR.R
\name{gibbsR}
\alias{gibbsR}
\title{A Gibbs sampler using R}
\usage{
gibbsR(N, thin)
}
\arguments{
\item{N}{the length of the chain}

\item{thin}{the number of Binomial(thin, y)}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsR(100,10)
plot(rnR[1,],type='l')
plot(rnR[2,],type='l')
}
}
