% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Matrix_Dimension.R
\name{Plot_Matrix_Dimension}
\alias{Plot_Matrix_Dimension}
\title{A function that Plot the number of dimensions from a segment to a segment}
\usage{
Plot_Matrix_Dimension(X = 100, Sstart = 1, Send = 100)
}
\arguments{
\item{X}{a matrix with  T*N size}

\item{Sstart}{the Start segment}

\item{Send}{the end segment, 1< Sstart<= Send < 100}
}
\value{
a number, the dimension of the matrix X
}
\description{
Plot the dimensions from a segment to a segment
}
\examples{
\dontrun{
X <- Plot_Matrix_Dimension()
}
}
