% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21049.R
\docType{package}
\name{StatComp21049}
\alias{StatComp21049}
\title{StatComp21049: A package for Calculate the phase space projection.}
\description{
The foo package provides three categories of important functions:
Embedding_delay,Downsample,Get_Matrix_Eigens,Get_Matrix_Dimension
}
\section{Foo functions}{

The foo functions…
}

