% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_Elapsed}
\alias{time_Elapsed}
\title{Example data of the time series for a worm}
\format{
A Matrix with 43288*100.
}
\source{
\url{http://www.wenlab.org/}
}
\usage{
data(time_Elapsed)
}
\description{
The unit is seconds.
}
\seealso{
\link{curve_data}
}
\keyword{datasets}
