% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w}
\alias{w}
\title{Example Projection matrix for a worm}
\format{
A Matrix with 60*7
}
\source{
\url{http://www.wenlab.org/}
}
\usage{
data(w)
}
\description{
Example Projection matrix for a worm
}
\seealso{
\link{curve_data},\link{time_Elapsed}
}
\keyword{datasets}
