% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.R
\name{Downsample}
\alias{Downsample}
\title{A function to downsample the raw data}
\usage{
Downsample(timeseq, dataseq, dft = 16)
}
\arguments{
\item{timeseq}{a vector with T*1 size
the unit is (s)}

\item{dataseq}{a matrix with T*N size}

\item{dft}{downsample to dfz (Hz)}
}
\value{
a list,contain $time_downsampled, $data_downsampled
}
\description{
Downsample the time sequence and data
}
\examples{
\dontrun{
data("curve_data","time_Elapsed")
X <- Plot_Matrix_Dimension(curve_data,time_Elapsed,16)
time_downsample
}
}
