% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21049.R
\docType{package}
\name{StatComp21049}
\alias{StatComp21049}
\title{StatComp21049: A package for Calculate the phase space projection.}
\description{
StatComp21049 is a R package developed to present the homework of statistic
computing course by author StatComp21049 and reconstruct the Phase space for worm behavior.
}
\details{
The StatComp21049 package provides three categories of important functions:
Embedding_delay,Downsample,Get_Matrix_Eigens,Get_Matrix_Dimension

You can You can visit the online webpage to view the help documentation from
\url{https://rongkangxiong.github.io/StatComp21049}
}
\section{StatComp21049 functions}{

\code{\link{Embedding_delay}} : Function to Embedding and delay the matrix after PCA projection.
\code{\link{Downsample}} : Function to Downsample data.
\code{\link{Get_Matrix_Eigens}} : Function to get eigenvalue and eigenvector.
\code{\link{Get_Matrix_Dimension}} : Get the matrix real dimension.

You can visit my blog:
\url{https://neuralife.cn}.

You can email me from \email{earth@mail.ustc.edu.cn}.
}

